FWF-4277: [Feature] Moved draft public APIs to public_endpoints file
